added more test cases and modified filman to pass test

+++ src/fileman.py
@@ -13,24 +13,36 @@ class FileMan(object):
 		reads a file and return content in a list
 		returns False is path is not a file
 		"""
-		if not Util.isfile(self.file_location):
-			Util.printline('make this file is in data folder')
+		if not self.validate():
 			return False
 
 		data = []
 		with open(self.file_location,'r') as file:
 			for line in file:
-				data.append(line)
+				data.append(line[:-1])
 			return data
 
 	def write(self,data):
+		"""appends content to a file"""
 		with open(self.file_location,'a+') as file:
 			file.read()
-			file.write(data)
+			file.write(data+'\n')
 
 	def replace(self,data):
+		"""replaces content of a file"""
 		with open(self.file_location,'w') as file:
-			file.write(data)
+			file.write(data+'\n')
 
 	def remove(self):
+		"""deletes file"""
+		if not self.validate():
+			return False
+
 		return os.remove(self.file_location)
+
+	def validate(self):
+		"""validates the existence of a give file location"""
+		if not Util.isfile(self.file_location):
+			Util.printline('make this file is in data folder')
+			return False
+		return True
